move Level, Placement, Space back to upper package
