+++ src/test/java/com/github/skjolberg/packing/Visualizer.java
@@ -1,8 +1,5 @@
 package com.github.skjolberg.packing;
 
-import com.github.skjolberg.packing.impl.Level;
-import com.github.skjolberg.packing.impl.Placement;
-import com.github.skjolberg.packing.impl.Space;
 import com.indvd00m.ascii.render.Render;
 import com.indvd00m.ascii.render.api.ICanvas;
 import com.indvd00m.ascii.render.api.IContextBuilder;
